+++ python/pathway/xpacks/llm/vector_store.py
@@ -9,15 +9,14 @@ HTTP requests.
 The client queries the server and returns matching documents.
 """
 
-import json
 import logging
 import threading
 import warnings
 from collections.abc import Callable, Coroutine
 from typing import TYPE_CHECKING, Iterable, TypeAlias, cast
+from warnings import warn
 
 import jmespath
-import requests
 
 import pathway as pw
 import pathway.xpacks.llm.parsers
@@ -27,6 +26,7 @@ from pathway.stdlib.indexing import default_usearch_knn_document_index
 from pathway.stdlib.indexing.data_index import _SCORE, DataIndex
 from pathway.stdlib.ml.classifiers import _knn_lsh
 from pathway.xpacks.llm._utils import _coerce_sync
+from pathway.xpacks.llm.document_store import DocumentStoreClient
 
 from ._utils import _unwrap_udf
 
@@ -648,11 +648,11 @@ class SlidesVectorStoreServer(VectorStoreServer):
         return self.inputs_query(parse_docs_queries)
 
 
-class VectorStoreClient:
+class VectorStoreClient(DocumentStoreClient):
     """
     A client you can use to query VectorStoreServer.
 
-    Please provide either the `url`, or `host` and `port`.
+    Please provide either the ``"url"``, or ``"host"`` and ``"port"``.
 
     Args:
         host: host on which `VectorStoreServer </developers/api-docs/pathway-xpacks-llm/vectorstore#pathway.xpacks.llm.vector_store.VectorStoreServer>`_ listens
@@ -661,109 +661,10 @@ class VectorStoreClient:
         timeout: timeout for the post requests in seconds
     """  # noqa
 
-    def __init__(
-        self,
-        host: str | None = None,
-        port: int | None = None,
-        url: str | None = None,
-        timeout: int | None = 15,
-        additional_headers: dict | None = None,
-    ):
-        err = "Either (`host` and `port`) or `url` must be provided, but not both."
-        if url is not None:
-            if host or port:
-                raise ValueError(err)
-            self.url = url
-        else:
-            if host is None:
-                raise ValueError(err)
-            port = port or 80
-            self.url = f"http://{host}:{port}"
-
-        self.timeout = timeout
-        self.additional_headers = additional_headers or {}
-
-    def query(
-        self,
-        query: str,
-        k: int = 3,
-        metadata_filter: str | None = None,
-        filepath_globpattern: str | None = None,
-    ) -> list[dict]:
-        """
-        Perform a query to the vector store and fetch results.
-
-        Args:
-            query:
-            k: number of documents to be returned
-            metadata_filter: optional string representing the metadata filtering query
-                in the JMESPath format. The search will happen only for documents
-                satisfying this filtering.
-            filepath_globpattern: optional glob pattern specifying which documents
-                will be searched for this query.
-        """
-
-        data = {"query": query, "k": k}
-        if metadata_filter is not None:
-            data["metadata_filter"] = metadata_filter
-        if filepath_globpattern is not None:
-            data["filepath_globpattern"] = filepath_globpattern
-        url = self.url + "/v1/retrieve"
-        response = requests.post(
-            url,
-            data=json.dumps(data),
-            headers=self._get_request_headers(),
-            timeout=self.timeout,
+    def __init__(self, *args, **kwargs):
+        warn(
+            "`VectorStoreClient` is being deprecated, please use `DocumentStoreClient` instead. "
+            "You can import with the following: `from pathway.xpacks.llm.document_store import DocumentStoreClient`",
+            DeprecationWarning,
         )
-
-        responses = response.json()
-        return sorted(responses, key=lambda x: x["dist"])
-
-    # Make an alias
-    __call__ = query
-
-    def get_vectorstore_statistics(self):
-        """Fetch basic statistics about the vector store."""
-
-        url = self.url + "/v1/statistics"
-        response = requests.post(
-            url,
-            json={},
-            headers=self._get_request_headers(),
-            timeout=self.timeout,
-        )
-        responses = response.json()
-        return responses
-
-    def get_input_files(
-        self,
-        metadata_filter: str | None = None,
-        filepath_globpattern: str | None = None,
-    ):
-        """
-        Fetch information on documents in the the vector store.
-
-        Args:
-            metadata_filter: optional string representing the metadata filtering query
-                in the JMESPath format. The search will happen only for documents
-                satisfying this filtering.
-            filepath_globpattern: optional glob pattern specifying which documents
-                will be searched for this query.
-        """
-        url = self.url + "/v1/inputs"
-        response = requests.post(
-            url,
-            json={
-                "metadata_filter": metadata_filter,
-                "filepath_globpattern": filepath_globpattern,
-            },
-            headers=self._get_request_headers(),
-            timeout=self.timeout,
-        )
-        responses = response.json()
-        return responses
-
-    def _get_request_headers(self):
-        request_headers = {"Content-Type": "application/json"}
-        request_headers.update(self.additional_headers)
-        return request_headers
+        super().__init__(*args, **kwargs)